live/trader/requirements.txt
Rx==3.0.0a0
krakenex==2.1.0
websocket-client==0.55.0
bitfinex-v2==1.1.5
